beberapa perubahan pada tabel barang

+++ database/migrations/2025_08_12_052221_create_barangs_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->string('nama_barang');
             $table->integer('qty');
-            $table->double('harga');
+            $table->decimal('harga', 10, 2);
             $table->timestamps();
         });
     }
